libtinyiiod: Use “stdint” types
Necessary so that when we use it with ad9361, cppcheck won't complain,
otherwise casts between types need to be used.
Add "compat.h" that will implement stdint types and
a smaller version of stdlib.
This is meant to be used with smaller uC, when the compiler
doesn’t come with “stdlib”.
Further implementation will be necessary, to be completely usable.
Functions like “malloc”, “free”, “snprintf” need to be implemented.
A cmake option was added, to select the smaller version of stdlib.

+++ tinyiiod-private.h
@@ -22,12 +22,12 @@
 
 char tinyiiod_read_char(struct tinyiiod *iiod);
 void tinyiiod_read(struct tinyiiod *iiod, char *buf, size_t len);
-ssize_t tinyiiod_read_line(struct tinyiiod *iiod, char *buf, size_t len);
+int32_t tinyiiod_read_line(struct tinyiiod *iiod, char *buf, size_t len);
 
 void tinyiiod_write_char(struct tinyiiod *iiod, char c);
 void tinyiiod_write(struct tinyiiod *iiod, const char *data, size_t len);
 void tinyiiod_write_string(struct tinyiiod *iiod, const char *str);
-void tinyiiod_write_value(struct tinyiiod *iiod, int value);
+void tinyiiod_write_value(struct tinyiiod *iiod, int32_t value);
 
 void tinyiiod_write_xml(struct tinyiiod *iiod);
 
@@ -45,6 +45,6 @@ void tinyiiod_do_close(struct tinyiiod *iiod, const char *device);
 void tinyiiod_do_readbuf(struct tinyiiod *iiod,
 			 const char *device, size_t bytes_count);
 
-int tinyiiod_parse_string(struct tinyiiod *iiod, char *str);
+int32_t tinyiiod_parse_string(struct tinyiiod *iiod, char *str);
 
 #endif /* TINYIIOD_PRIVATE_H */
